variant template seeder and factory added

+++ database/migrations/2025_12_16_110231_create_materials_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
 
             $table->string('surface');
             $table->string('composition');
-            $table->float('percentage');
+            $table->decimal('percentage', 5);
 
             $table->uuid('category_id');
             $table->foreign('category_id')
